ahoy/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
groq==0.11.0
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
cachetools==5.5.0
slowapi==0.1.9
